+++ NEWS.md
@@ -7,6 +7,7 @@
 - Leaf area, foliar biomass and fuels of shrubs and herbs depend on the leaf area of plants above
 - Survival model based on basal area available as an alternative to fixed mortality baseline rates
 - New control thresholds 'minimumShrubCohortCover' and 'minimumTreeCohortDensity' for cohort removal
+- Tree bark thickness parameters added to SpParamsMED
 
 # Version 3.1.1
 - Fire severity implemented